UPD: Merged revision(s) 6959, 6968 from trunk
git-svn-id: https://svn.code.sf.net/p/doublecmd/code@6969 10bb5c37-c6d6-4c74-9a2a-30fc9b6e22f7

+++ branches/0.7/src/ucolumns.pas
@@ -1137,7 +1137,7 @@ begin
   ANode := ANode.FindNode('ColumnsSets');
   if Assigned(ANode) then
   begin
-    DefaultTitleHash := AConfig.GetAttr(ANode, 'DefaultTitleHash', 0);
+    DefaultTitleHash := AConfig.GetAttr(ANode, 'DefaultTitleHash', Int64(0));
     ANode := ANode.FirstChild;
     while Assigned(ANode) do
     begin
@@ -1165,7 +1165,7 @@ var
 begin
   ANode := AConfig.FindNode(ANode, 'ColumnsSets', True);
   AConfig.ClearNode(ANode);
-  AConfig.SetAttr(ANode, 'DefaultTitleHash', DefaultTitleHash);
+  AConfig.SetAttr(ANode, 'DefaultTitleHash', Int64(DefaultTitleHash));
   for I := 0 to FSet.Count - 1 do
   begin
     SubNode := AConfig.AddNode(ANode, 'ColumnsSet');
